+++ src/controllers/TemplatesController.php
@@ -40,7 +40,6 @@ class TemplatesController extends Controller
     public $allowAnonymous = [
         'offline' => self::ALLOW_ANONYMOUS_LIVE | self::ALLOW_ANONYMOUS_OFFLINE,
         'manual-update-notification' => self::ALLOW_ANONYMOUS_LIVE | self::ALLOW_ANONYMOUS_OFFLINE,
-        'config-sync-kickoff' => self::ALLOW_ANONYMOUS_LIVE | self::ALLOW_ANONYMOUS_OFFLINE,
         'requirements-check' => self::ALLOW_ANONYMOUS_LIVE | self::ALLOW_ANONYMOUS_OFFLINE,
         'render-error' => self::ALLOW_ANONYMOUS_LIVE | self::ALLOW_ANONYMOUS_OFFLINE,
     ];
@@ -120,16 +119,6 @@ class TemplatesController extends Controller
     public function actionManualUpdateNotification(): Response
     {
         return $this->renderTemplate('_special/dbupdate');
-    }
-
-    /**
-     * Renders the Project Config Sync kickoff template.
-     *
-     * @return Response
-     */
-    public function actionConfigSyncKickoff(): Response
-    {
-        return $this->renderTemplate('_special/configsync');
     }
 
     /**